+++ src/Either.ts
@@ -121,8 +121,8 @@ export function right<E = never, A = never>(a: A): Either<E, A> {
  *
  * @since 2.0.0
  */
-export function fromNullable<E>(f: () => E): <A>(a: A) => Either<E, NonNullable<A>> {
-  return <A>(a: A) => (a == null ? left(f()) : right(a as NonNullable<A>))
+export function fromNullable<E>(e: () => E): <A>(a: A) => Either<E, NonNullable<A>> {
+  return <A>(a: A) => (a == null ? left(e()) : right(a as NonNullable<A>))
 }
 
 /**