+++ src/TaskThese.ts
@@ -120,7 +120,7 @@ export function getMonad<E>(S: Semigroup<E>): Monad2C<URI, E> & MonadTask2C<URI,
 /**
  * @since 2.4.0
  */
-export function toTuple<E, A>(e: E, a: A): (fa: TaskThese<E, A>) => Task<[E, A]> {
+export function toTuple<E, A>(e: () => E, a: () => A): (fa: TaskThese<E, A>) => Task<[E, A]> {
   return (fa) => T.toTuple(fa, e, a)
 }
 /* tslint:enable:readonly-array */